requirements.txt
fastapi
openai
python-dotenv
python-docx
uvicorn[standard]